+++ Changelog.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-*empty*
+### Added
+- Add ability to print out location of config directory
 
 ## [1.3.0] - 2022-05-18
 ### Added